Remove UpdatablePageSource and clients
