+++ core/trino-main/src/main/java/io/trino/operator/MergeWriterOperator.java
@@ -13,27 +13,42 @@
  */
 package io.trino.operator;
 
+import com.google.common.collect.ImmutableList;
+import com.google.common.util.concurrent.Futures;
+import com.google.common.util.concurrent.ListenableFuture;
+import io.airlift.slice.Slice;
 import io.trino.Session;
 import io.trino.spi.Page;
+import io.trino.spi.PageBuilder;
 import io.trino.spi.block.Block;
+import io.trino.spi.block.BlockBuilder;
 import io.trino.spi.connector.ConnectorMergeSink;
+import io.trino.spi.type.Type;
 import io.trino.split.PageSinkId;
 import io.trino.split.PageSinkManager;
 import io.trino.sql.planner.plan.PlanNodeId;
 import io.trino.sql.planner.plan.TableWriterNode.MergeTarget;
 
+import java.util.Collection;
+import java.util.List;
 import java.util.function.Function;
 import java.util.stream.IntStream;
 
 import static com.google.common.base.Preconditions.checkState;
+import static com.google.common.util.concurrent.MoreExecutors.directExecutor;
+import static io.airlift.concurrent.MoreFutures.getFutureValue;
 import static io.airlift.concurrent.MoreFutures.toListenableFuture;
 import static io.trino.spi.connector.ConnectorMergeSink.IGNORED_OPERATION_NUMBER;
+import static io.trino.spi.type.BigintType.BIGINT;
 import static io.trino.spi.type.TinyintType.TINYINT;
+import static io.trino.spi.type.VarbinaryType.VARBINARY;
 import static java.util.Objects.requireNonNull;
 
 public class MergeWriterOperator
-        extends AbstractRowChangeOperator
+        implements Operator
 {
+    private static final List<Type> TYPES = ImmutableList.of(BIGINT, VARBINARY);
+
     public static class MergeWriterOperatorFactory
             implements OperatorFactory
     {
@@ -78,15 +93,43 @@ public class MergeWriterOperator
         }
     }
 
-    private final ConnectorMergeSink mergeSink;
+    private enum State
+    {
+        RUNNING, FINISHING, FINISHED
+    }
 
+    private final OperatorContext operatorContext;
+    private State state = State.RUNNING;
+    private final ConnectorMergeSink mergeSink;
     private final Function<Page, Page> pagePreprocessor;
+    private ListenableFuture<Collection<Slice>> finishFuture;
+    private ListenableFuture<Void> blockedFutureView;
+    private long rowCount;
+    private boolean closed;
 
     public MergeWriterOperator(OperatorContext operatorContext, ConnectorMergeSink mergeSink, Function<Page, Page> pagePreprocessor)
     {
-        super(operatorContext);
+        this.operatorContext = requireNonNull(operatorContext, "operatorContext is null");
         this.mergeSink = requireNonNull(mergeSink, "mergeSink is null");
         this.pagePreprocessor = requireNonNull(pagePreprocessor, "pagePreprocessor is null");
+    }
+
+    @Override
+    public OperatorContext getOperatorContext()
+    {
+        return operatorContext;
+    }
+
+    @Override
+    public boolean isFinished()
+    {
+        return state == State.FINISHED;
+    }
+
+    @Override
+    public boolean needsInput()
+    {
+        return state == State.RUNNING;
     }
 
     @Override
@@ -119,17 +162,68 @@ public class MergeWriterOperator
     }
 
     @Override
+    public Page getOutput()
+    {
+        if ((state != State.FINISHING) || !finishFuture.isDone()) {
+            return null;
+        }
+        state = State.FINISHED;
+
+        Collection<Slice> fragments = getFutureValue(finishFuture);
+
+        // output page will only be constructed once,
+        // so a new PageBuilder is constructed (instead of using PageBuilder.reset)
+        PageBuilder page = new PageBuilder(fragments.size() + 1, TYPES);
+        BlockBuilder rowsBuilder = page.getBlockBuilder(0);
+        BlockBuilder fragmentBuilder = page.getBlockBuilder(1);
+
+        // write row count
+        page.declarePosition();
+        BIGINT.writeLong(rowsBuilder, rowCount);
+        fragmentBuilder.appendNull();
+
+        // write fragments
+        for (Slice fragment : fragments) {
+            page.declarePosition();
+            rowsBuilder.appendNull();
+            VARBINARY.writeSlice(fragmentBuilder, fragment);
+        }
+
+        return page.build();
+    }
+
+    @Override
     public void finish()
     {
         if (state == State.RUNNING) {
             state = State.FINISHING;
             finishFuture = toListenableFuture(mergeSink.finish());
+            blockedFutureView = asVoid(finishFuture);
         }
     }
 
-    @Override
-    protected void abort()
+    private static <T> ListenableFuture<Void> asVoid(ListenableFuture<T> future)
     {
-        mergeSink.abort();
+        return Futures.transform(future, v -> null, directExecutor());
+    }
+
+    @Override
+    public ListenableFuture<Void> isBlocked()
+    {
+        if (blockedFutureView == null) {
+            return NOT_BLOCKED;
+        }
+        return blockedFutureView;
+    }
+
+    @Override
+    public void close()
+    {
+        if (!closed) {
+            closed = true;
+            if (finishFuture != null) {
+                finishFuture.cancel(true);
+            }
+        }
     }
 }
