+++ core/trino-main/src/main/java/io/trino/operator/WorkProcessorPipelineSourceOperator.java
@@ -28,7 +28,6 @@ import io.trino.metadata.Split;
 import io.trino.operator.OperationTimer.OperationTiming;
 import io.trino.operator.WorkProcessor.ProcessState;
 import io.trino.spi.Page;
-import io.trino.spi.connector.UpdatablePageSource;
 import io.trino.spi.metrics.Metrics;
 import io.trino.spi.type.Type;
 import io.trino.sql.planner.LocalExecutionPlanner.OperatorFactoryWithTypes;
@@ -41,7 +40,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.Supplier;
 
 import static com.google.common.base.Preconditions.checkState;
 import static com.google.common.base.Throwables.throwIfUnchecked;
@@ -389,10 +387,10 @@ public class WorkProcessorPipelineSourceOperator
     }
 
     @Override
-    public Supplier<Optional<UpdatablePageSource>> addSplit(Split split)
+    public void addSplit(Split split)
     {
         if (sourceOperator == null) {
-            return Optional::empty;
+            return;
         }
 
         Object splitInfo = split.getInfo();
@@ -402,8 +400,6 @@ public class WorkProcessorPipelineSourceOperator
 
         pendingSplits.add(split);
         blockedOnSplits.set(null);
-
-        return sourceOperator.getUpdatablePageSourceSupplier();
     }
 
     @Override
