+++ core/trino-main/src/main/java/io/trino/operator/WorkProcessorSourceOperator.java
@@ -15,11 +15,7 @@ package io.trino.operator;
 
 import io.airlift.units.DataSize;
 import io.airlift.units.Duration;
-import io.trino.spi.connector.UpdatablePageSource;
 import io.trino.spi.metrics.Metrics;
-
-import java.util.Optional;
-import java.util.function.Supplier;
 
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
 
@@ -27,8 +23,6 @@ import static java.util.concurrent.TimeUnit.NANOSECONDS;
 public interface WorkProcessorSourceOperator
         extends WorkProcessorOperator
 {
-    Supplier<Optional<UpdatablePageSource>> getUpdatablePageSourceSupplier();
-
     default DataSize getPhysicalInputDataSize()
     {
         return DataSize.ofBytes(0);
