+++ core/trino-main/src/main/java/io/trino/operator/WorkProcessorSourceOperatorAdapter.java
@@ -20,14 +20,11 @@ import io.trino.Session;
 import io.trino.memory.context.MemoryTrackingContext;
 import io.trino.metadata.Split;
 import io.trino.spi.Page;
-import io.trino.spi.connector.UpdatablePageSource;
 import io.trino.spi.metrics.Metrics;
 import io.trino.sql.planner.plan.PlanNodeId;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
-import java.util.function.Supplier;
 
 import static io.trino.operator.WorkProcessor.ProcessState.blocked;
 import static io.trino.operator.WorkProcessor.ProcessState.finished;
@@ -95,10 +92,10 @@ public class WorkProcessorSourceOperatorAdapter
     }
 
     @Override
-    public Supplier<Optional<UpdatablePageSource>> addSplit(Split split)
+    public void addSplit(Split split)
     {
         if (operatorFinishing) {
-            return Optional::empty;
+            return;
         }
 
         Object splitInfo = split.getInfo();
@@ -107,7 +104,6 @@ public class WorkProcessorSourceOperatorAdapter
         }
 
         splitBuffer.add(split);
-        return sourceOperator.getUpdatablePageSourceSupplier();
     }
 
     @Override
