+++ core/trino-main/src/test/java/io/trino/connector/MockConnectorPageSource.java
@@ -13,25 +13,18 @@
  */
 package io.trino.connector;
 
-import io.airlift.slice.Slice;
 import io.trino.spi.Page;
-import io.trino.spi.block.Block;
 import io.trino.spi.connector.ConnectorPageSource;
-import io.trino.spi.connector.UpdatablePageSource;
 import io.trino.spi.metrics.Metrics;
 
 import java.io.IOException;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
 import java.util.OptionalLong;
 import java.util.concurrent.CompletableFuture;
 
 import static java.util.Objects.requireNonNull;
-import static java.util.concurrent.CompletableFuture.completedFuture;
 
 public class MockConnectorPageSource
-        implements UpdatablePageSource
+        implements ConnectorPageSource
 {
     private final ConnectorPageSource delegate;
     private final Metrics metrics;
@@ -95,23 +88,5 @@ public class MockConnectorPageSource
     public Metrics getMetrics()
     {
         return delegate.getMetrics().mergeWith(metrics);
-    }
-
-    @Override
-    public void deleteRows(Block rowIds) {}
-
-    @Override
-    public void updateRows(Page page, List<Integer> columnValueAndRowIdChannels) {}
-
-    @Override
-    public CompletableFuture<Collection<Slice>> finish()
-    {
-        return completedFuture(Collections.emptyList());
-    }
-
-    @Override
-    public void abort()
-    {
-        UpdatablePageSource.super.abort();
     }
 }
