+++ core/trino-spi/src/main/java/io/trino/spi/connector/UpdatablePageSource.java
@@ -21,9 +21,11 @@ import java.util.Collection;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
+@Deprecated // This interface has been removed from the API
 public interface UpdatablePageSource
         extends ConnectorPageSource
 {
+    @Deprecated // This method has been removed from the API
     default void deleteRows(Block rowIds)
     {
         throw new UnsupportedOperationException("This connector does not support row-level delete");
@@ -32,15 +34,18 @@ public interface UpdatablePageSource
     /**
      * Write updated rows to the PageSource.
      * @param page Contains values for all updated columns, as well as the $row_id column. The order of these Blocks can be derived from columnValueAndRowIdChannels.
-     * @param columnValueAndRowIdChannels The index of this list matches the index columns have in updatedColumns parameter of {@link ConnectorMetadata#beginUpdate}
+     * @param columnValueAndRowIdChannels The index of this list matches the index columns have in updatedColumns parameter of beginUpdate
      * The value at each index is the channel number in the given page. The last element of this list is always the channel number for the $row_id column.
      */
+    @Deprecated // This method has been removed from the API
     default void updateRows(Page page, List<Integer> columnValueAndRowIdChannels)
     {
         throw new UnsupportedOperationException("This connector does not support row update");
     }
 
+    @Deprecated // This method has been removed from the API
     CompletableFuture<Collection<Slice>> finish();
 
+    @Deprecated // This method has been removed from the API
     default void abort() {}
 }
