+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/HiveColumnHandle.java
@@ -22,12 +22,10 @@ import io.trino.spi.connector.ColumnMetadata;
 import io.trino.spi.type.Type;
 import org.openjdk.jol.info.ClassLayout;
 
-import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
 import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.collect.ImmutableList.toImmutableList;
 import static io.airlift.slice.SizeOf.estimatedSizeOf;
 import static io.airlift.slice.SizeOf.sizeOf;
 import static io.trino.plugin.hive.HiveColumnHandle.ColumnType.PARTITION_KEY;
@@ -36,7 +34,6 @@ import static io.trino.plugin.hive.HiveType.HIVE_INT;
 import static io.trino.plugin.hive.HiveType.HIVE_LONG;
 import static io.trino.plugin.hive.HiveType.HIVE_STRING;
 import static io.trino.plugin.hive.HiveType.toHiveType;
-import static io.trino.plugin.hive.HiveUpdateProcessor.getRowIdColumnHandleForNonUpdatedColumns;
 import static io.trino.plugin.hive.acid.AcidSchema.ACID_ROW_ID_ROW_TYPE;
 import static io.trino.spi.type.BigintType.BIGINT;
 import static io.trino.spi.type.IntegerType.INTEGER;
@@ -259,20 +256,6 @@ public class HiveColumnHandle
     public Column toMetastoreColumn()
     {
         return new Column(name, getHiveType(), comment);
-    }
-
-    public static HiveColumnHandle getDeleteRowIdColumnHandle()
-    {
-        return createBaseColumn(UPDATE_ROW_ID_COLUMN_NAME, UPDATE_ROW_ID_COLUMN_INDEX, toHiveType(ACID_ROW_ID_ROW_TYPE), ACID_ROW_ID_ROW_TYPE, SYNTHESIZED, Optional.empty());
-    }
-
-    public static HiveColumnHandle updateRowIdColumnHandle(List<HiveColumnHandle> columnHandles, List<ColumnHandle> updatedColumns)
-    {
-        requireNonNull(updatedColumns, "updatedColumns is null");
-        List<HiveColumnHandle> nonUpdatedColumnHandles = columnHandles.stream()
-                .filter(column -> !column.isPartitionKey() && !column.isHidden() && !updatedColumns.contains(column))
-                .collect(toImmutableList());
-        return getRowIdColumnHandleForNonUpdatedColumns(nonUpdatedColumnHandles);
     }
 
     public static HiveColumnHandle mergeRowIdColumnHandle()
