+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/HiveTableHandle.java
@@ -225,29 +225,6 @@ public class HiveTableHandle
                 maxScannedFileSize);
     }
 
-    public HiveTableHandle withUpdateProcessor(AcidTransaction transaction, HiveUpdateProcessor updateProcessor)
-    {
-        requireNonNull(updateProcessor, "updateProcessor is null");
-        return new HiveTableHandle(
-                schemaName,
-                tableName,
-                tableParameters,
-                partitionColumns,
-                dataColumns,
-                partitionNames,
-                partitions,
-                compactEffectivePredicate,
-                enforcedConstraint,
-                bucketHandle,
-                bucketFilter,
-                analyzePartitionValues,
-                constraintColumns,
-                projectedColumns,
-                transaction,
-                recordScannedFiles,
-                maxScannedFileSize);
-    }
-
     public HiveTableHandle withProjectedColumns(Set<ColumnHandle> projectedColumns)
     {
         return new HiveTableHandle(
@@ -424,27 +401,9 @@ public class HiveTableHandle
     }
 
     @JsonIgnore
-    public boolean isAcidDelete()
-    {
-        return transaction.isDelete();
-    }
-
-    @JsonIgnore
-    public boolean isAcidUpdate()
-    {
-        return transaction.isUpdate();
-    }
-
-    @JsonIgnore
     public boolean isAcidMerge()
     {
         return transaction.isMerge();
-    }
-
-    @JsonIgnore
-    public Optional<HiveUpdateProcessor> getUpdateProcessor()
-    {
-        return transaction.getUpdateProcessor();
     }
 
     @JsonIgnore
