+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/orc/OrcFileWriterFactory.java
@@ -29,7 +29,6 @@ import io.trino.orc.metadata.CompressionKind;
 import io.trino.plugin.hive.FileFormatDataSourceStats;
 import io.trino.plugin.hive.FileWriter;
 import io.trino.plugin.hive.HiveFileWriterFactory;
-import io.trino.plugin.hive.HiveType;
 import io.trino.plugin.hive.NodeVersion;
 import io.trino.plugin.hive.WriterKind;
 import io.trino.plugin.hive.acid.AcidTransaction;
@@ -68,7 +67,6 @@ import static io.trino.plugin.hive.HiveSessionProperties.getOrcOptimizedWriterVa
 import static io.trino.plugin.hive.HiveSessionProperties.getOrcStringStatisticsLimit;
 import static io.trino.plugin.hive.HiveSessionProperties.getTimestampPrecision;
 import static io.trino.plugin.hive.HiveSessionProperties.isOrcOptimizedWriterValidate;
-import static io.trino.plugin.hive.HiveType.toHiveType;
 import static io.trino.plugin.hive.acid.AcidSchema.ACID_COLUMN_NAMES;
 import static io.trino.plugin.hive.acid.AcidSchema.createAcidColumnPrestoTypes;
 import static io.trino.plugin.hive.acid.AcidSchema.createRowType;
@@ -155,11 +153,6 @@ public class OrcFileWriterFactory
         int[] fileInputColumnIndexes = fileColumnNames.stream()
                 .mapToInt(inputColumnNames::indexOf)
                 .toArray();
-        if (transaction.isAcidDeleteOperation(writerKind)) {
-            // For delete, set the "row" column to -1
-            fileInputColumnIndexes[fileInputColumnIndexes.length - 1] = -1;
-        }
-
         try {
             TrinoFileSystem fileSystem = fileSystemFactory.create(session);
             String stringPath = path.toString();
@@ -224,17 +217,6 @@ public class OrcFileWriterFactory
         catch (IOException e) {
             throw new TrinoException(HIVE_WRITER_OPEN_ERROR, "Error creating ORC file", e);
         }
-    }
-
-    public static HiveType createHiveRowType(Properties schema, TypeManager typeManager, ConnectorSession session)
-    {
-        List<String> dataColumnNames = getColumnNames(schema);
-        List<Type> dataColumnTypes = getColumnTypes(schema).stream()
-                .map(hiveType -> hiveType.getType(typeManager, getTimestampPrecision(session)))
-                .collect(toList());
-        Type dataRowType = createRowType(dataColumnNames, dataColumnTypes);
-        Type acidRowType = createRowType(ACID_COLUMN_NAMES, createAcidColumnPrestoTypes(dataRowType));
-        return toHiveType(acidRowType);
     }
 
     public static OrcDataSink createOrcDataSink(TrinoFileSystem fileSystem, String path)
